Retain namespace reference until orchestrator terminates.
An anomaly was noted with internally modeled namespaces when the
orchestrator was instructed to output a security model for a
namespace invocation.  Only a subset of the trajectory entries
were being output when trajectory output was requested.

The root cause of this issue was secondary to the fact that the
context describing the security modeling namespace was being
released when the last process in the namespace terminated.  This
is incorrect behavior in the new architecture where the
orchestrator has a direct reference to the namespace that is
being modeled.

The trust orchestrator for an externally modeled namespace has to
read the model paramters AFTER the last process in the modeling
namespace has terminated.  Secondary to this the orchestrator
process has to also hold a reference to the context describing
the namespace, so the context is not destroyed before the
orchestrator has the ability to read all of the event in the
model.

The namespace creation code was updated to take a reference to
the security namespace context as part of the security namespace
creation process.  This insurances that a reference is held for
the orchestrator process.

The tsem_task_free() security event handler was updated to check
to see if the exiting task has a non-NULL tma_context value.  If
so the reference for the orchestrator process is released.