Update control plane handling of security namespace identifier.
The changes that involve having the orchestrator process have
access to the characteristics of the security modeling namespace
required a change to the control plane file that returns the
identifier of the namespace.

The tsem_tma_context() function was used to return a pointer to
the subordinate namespace definition.

The CAP_MAC_ADMIN capability is now required in order to open the
control plane file.

From a userspace perspective the most important change was to
return ENOENT if the subordinate process creating the namespace
has not yet registered a pointer to the security modeling context
in the task control structure of the orchestrator process.  This
provides a scheme to allow the the trust orchestrator process to
poll for whether or not the namespace has been established.

In the case of internally modeled namespaces this provides a
method to block the orchestrator from configuring the namespace
until the namespace has been established.  The userspace trust
orchestrator blocks on the open of the identifier and when the
open succeeds moves forward with configuration of the namespace.