Make function for retriving TMA context sub-system visible.
A regression has been noted in the orchestrator that manages
internally modeled namespaces.  A sealed model is generating both
trajectory events and forensic events.

The issue appears to be secondary to the fact that the routines
that load model coefficients are referencing the incorrect model.
A factor being caused by the fact that the context used to locate
the model is the context of the process loading the coefficients
rather than the context of the workload being modeled.

The inline function that returns the TMA model was moved to
tsem.h since both the fs.c and model.c files need to access it.

In addition the function was fixed since it was calling the
tsem_tma_context() function with the 'current' task rather than
the task that was supplied as an argument to the inline function.

All of the calls in the fs.c file were converted to
tsem_tma_model rather than the file specific tma_model function.