Use correct context for model configuration.
An anomaly was noted with the new orchestrator.  Internally
modeled namespaces that were running with a sealed model were
generating both trajectory and forensics events.  In a sealed
model that does not experience an unmodeled event there should be
no trajectory or forensic events.

The following functions were converted to use the
tsem_tma_context() function to reference the correct modeling
context when called by the trust orchestrator.

tsem_model_load_point()
tsem_model_load_base()
tsem_model_load_pseudonym()

The tsem_model_load_point() function was referencing the
following two statically scoped functions:

have_point()
add_event_point()

These two functions were modified to take as their first argument
the security modeling context that is to be used for the
invocation of the function.

This allows the functions, when called by the trust orchestrator
to use the context of the namespace that the orchestrator is
supervising.  Tasks running in the context of the namespace use
the context of the process generating the event being modeled as
the argument to these functions.