Record events by untrusted processes as forensic events.
Work with developing models indicated that in kernel modeling was
operating in a manner that violated least surprise with respect
to when model violations were occurring.  Usually only a single
forensics event was registered, followed by multiple warning
messages in the kernel logs.

The warning messages for security events that occur when a
process is running in untrusted status but these events are not
registered as forensics events in the in kernel TMA.

In order to address this situation the following function was
added with sub-system global scope:

tsem_model_add_violation()

This function takes an event description structure and registers
it as a forensics event with an associated coefficient.

This allows updates to a security model to be derived directly
from the security state coefficients that are read from the
/sys/kernel/security/tsem/internal_tma/model0/forensics_coefficients
control plane file.