Replace log exports with security event description exports.
Previously when an untrusted process generated security events a
log record describing the process and the event were exported to
userspace.  In order to be able to fold forensic events back into
a model this needed to be changed so a record of event
descriptions are available so the events can be incorporated into
an updated model.

The following function:

tsem_export_log()

Was renamed to be as follows:

tsem_export_violation()

To more correctly describe what the purpose of the function is.

The tsem_export_violation() was modified to initialize the event
description structure that was passed to it.  The event
description is then exported to userspace.

The tsem_export_show() function was modified so that the
LOG_EVENT action writes the description of the security event to
the export control plane file.