Rename model events that handle security coefficients.
In the original nomenclature security state coefficients were
referred to as 'points'.  This is consistent with what current
nomenclature has evolved to.

Rename the following functions that handle security coefficients:

add_event_point()
have_point()

To the following:

add_coefficient()
have_coefficient()

The following functions were also misnamed:

add_trajectory_point()
add_forensic_point()

Since they add security event descriptions, not security state
coefficients.

The latter two functions simply insert the security state
description into its appropriate list and do have the need to
generate a return value so change the function to a void return
type and correct their call points to not expect that a return
value should be handled.

A correctness fix was also implemented since a failure to
allocate a security state coefficient resulted in the event not
being added to the model, and in the case for a forensic event,
the process was not set to an untrusted state.