Allow access to model information from root modeling namespace.
A regression was noted where it was not possible to read the
model information out of the internal TMA for the root modeling
namespace.

This was secondary to the ordering of the permissions checks in
the following function:

fs.c:can_access_fs()

The first check in that function was whether or not the the TMA
context was a non-NULL pointer.  In the case of processes running
in the root modeling namespace this value will always be null.

The check for whether or not the process requesting the file open
had the CAP_MAC_ADMIN capabilities was in front of this check.
The CAP_MAC_ADMIN check was modified to also require that the
process was running in the root security modeling namespace.

The tsem.c:tsem_init() function was modified to initialize the
tma_context member of the TSEM task control structure to have as
its value a pointer to the root security modeling context.