Implemented hashed coefficient lookups for internal models.
The initial internal Trusted Modeling Agent (TMA) implementation
was naive with respect to performance, particularly for the root
modeling namespace.

Every security event requires a search of the security state
coefficient list to determine if the coefficient for an event
already exists.  The coefficients are saved on a linked list for
the model, a fact that results in the search of that list being
an extremely performance critical path.

In the case of the root security modeling namespace this list can
be thousands or tens of thousands of entries long resulting in
significant latency for each security event.

The use of a cryptographic checksum as the generative function
for a security state coefficient provides a natural framework for
the implementation of a hash table to decrease the length of the
search chains.

This commit adds a new member, coeff_lists, to the tsem_model
structure.  This new member is an array of 256 list_head
structures.  The first byte of a security state coefficient is
used as an index into this array to determine the list that a
coefficient should be placed on and searched for.

The select_list() function was added to the model.c file in order
to encapsulate the selection of the correct list for a specific
coefficient.  An additional list structure was added to the
tsem_event_point, named lookup_list, that is used to enroll a
security state coefficient onto its appropriate hashed list.

Earlier nomenclature had used the term 'point' to specify a
security state coefficient.  This commit updates the members of
the tsem_model structure to use 'coeff' rather than 'point' in
the naming of the structure.  The majority of the size of this
commit is secondary to the renaming that this caused.

For simplicity purposes the original single linked list of
security state coefficients was retained in order to simplify
control plane access to the coefficients.  This allows the
control plane to provide a temporal ordering of when the security
state coefficients were generated.