Properly reference the root security context.
Testing with root security namespace modeling indicated a
regression when the control plane files were accessed.  Most
specifically a null pointer reference would be generated when the
control plane files were accessed.

This was caused by the fact that, be design, the TMA context was
not transferred to a subordinate task in the tsem_task_alloc()
handler.  This was by design in order to prevent any other
process, other than the trust orchestrator for a subordinate
security modeling namespace, from accessing or controlling the
namespace.

This has the negative effect of leaving the tsem_task->tma_context
pointer NULL for any process other than the init process.

The tma_context initialization was removed from the tsem_init()
function since that initialization was not useful.

The tsem_tma_context() inline function was modified to return
the root modeling namespace context if a process is in the root
namespace and it's tma_context is NULL.  This condition indicates
that the process accessing the control plane is in the root
modeling namespace but is not an orchestrator for a subordinate
modeling namespace.

This change required a modification in the logic that issues a
'put' against the namespace context by a trust orchestrator.  In
order for the 'put' to be issued the security modeling context of
the calling must not match the tma_context.  This is the
definition of the condition that a trust orchestrator has in
place.