Drop namespace specific export files.
Userspace development of the Quixote tooling has suggested that
there is little value in having separate control plane files for
security modeling namespaces that are exporting security event
descriptions to user.  Having separate control files also
marginally complicates the userspace implementations.

This commit drops the /sys/kernel/security/tsem/external_tma
directory and replaces it with the following control plane file:

/sys/kernel/security/tsem/export.

A userspace trust orchestrator only needs to open this file in
order to get an event stream for a security modeling namespace
that it is orchestrating.